+++ scripts/test/SysConfig/XML.t
@@ -168,13 +168,13 @@ my @Tests = (
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
+                        },
                     ],
                     'Description' => [
                         {
                             'Content'      => 'Test 1.',
                             'Translatable' => '1'
-                        }
+                        },
                     ],
                     'Name'     => 'Test1',
                     'Required' => '1',
@@ -185,10 +185,10 @@ my @Tests = (
                                     'ValueRegex' => '.*',
                                     'Content'    => '123',
                                     'ValueType'  => 'String'
-                                }
-                                ]
-                        }
-                        ]
+                                },
+                            ],
+                        },
+                    ],
                 },
                 'XMLContentRaw' => '<Setting Name="Test1" Required="1" Valid="1">
         <Description Translatable="1">Test 1.</Description>
@@ -212,7 +212,7 @@ my @Tests = (
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
+                        },
                     ],
                     'Required' => '1',
                     'Value'    => [
@@ -221,9 +221,9 @@ my @Tests = (
                                 {
                                     'ValueType' => 'File',
                                     'Content'   => '/usr/bin/gpg'
-                                }
-                                ]
-                        }
+                                },
+                            ],
+                        },
                     ],
                     'Valid' => '1'
                 },
@@ -274,13 +274,13 @@ my @Tests = (
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
+                        },
                     ],
                     'Description' => [
                         {
                             'Translatable' => '1',
                             'Content'      => 'Test 1.'
-                        }
+                        },
                     ],
                     'Required' => '1',
                     'Value'    => [
@@ -291,14 +291,14 @@ my @Tests = (
                                         "\x{e4}\x{eb}\x{ef}\x{f6}\x{fc}\x{c4}\x{cb}\x{cf}\x{d6}\x{dc}\x{e1}\x{e9}\x{ed}\x{f3}\x{fa}\x{c1}\x{c9}\x{cd}\x{d3}\x{da}\x{f1}\x{d1}\x{20ac}\x{438}\x{441}\x{df}",
                                     'ValueRegex' => '.*',
                                     'ValueType'  => 'String'
-                                }
-                                ]
-                        }
+                                },
+                            ],
+                        },
                     ],
                     'Name' =>
                         "\x{e4}\x{eb}\x{ef}\x{f6}\x{fc}\x{c4}\x{cb}\x{cf}\x{d6}\x{dc}\x{e1}\x{e9}\x{ed}\x{f3}\x{fa}\x{c1}\x{c9}\x{cd}\x{d3}\x{da}\x{f1}\x{d1}\x{20ac}\x{438}\x{441}\x{df}",
                     'Valid' => '1'
-                    }
+                },
             },
             {
                 'XMLFilename'   => undef,
@@ -316,9 +316,9 @@ my @Tests = (
                                 {
                                     'ValueType' => 'File',
                                     'Content'   => '/usr/bin/gpg'
-                                }
-                                ]
-                        }
+                                },
+                            ],
+                        },
                     ],
                     'Valid'       => '1',
                     'Name'        => 'Test2',
@@ -326,15 +326,15 @@ my @Tests = (
                         {
                             'Content'      => 'Test 2.',
                             'Translatable' => '1'
-                        }
+                        },
                     ],
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
+                        },
                     ],
                     'Required' => '1'
-                    }
+                },
             }
         ],
     },
@@ -370,7 +370,7 @@ my @Tests = (
                         {
                             'Translatable' => '1',
                             'Content'      => 'Test 1.'
-                        }
+                        },
                     ],
                     'Valid' => '1',
                     'Value' => [
@@ -380,15 +380,15 @@ my @Tests = (
                                     'Content'    => '123',
                                     'ValueType'  => 'String',
                                     'ValueRegex' => '.*'
-                                }
-                                ]
-                        }
+                                },
+                            ],
+                        },
                     ],
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
-                        ]
+                        },
+                    ],
                 },
                 'XMLContentRaw' => '<Setting Name="Test1" Required="1" Valid="1">
         <Description Translatable="1">Test 1.</Description>
@@ -404,14 +404,14 @@ my @Tests = (
                         {
                             'Content'      => 'Test 2.',
                             'Translatable' => '1'
-                        }
+                        },
                     ],
                     'Name'       => 'Test2',
                     'Required'   => '1',
                     'Navigation' => [
                         {
                             'Content' => 'Core::Ticket'
-                        }
+                        },
                     ],
                     'Value' => [
                         {
@@ -419,9 +419,9 @@ my @Tests = (
                                 {
                                     'ValueType' => 'File',
                                     'Content'   => '/usr/bin/gpg'
-                                }
-                                ]
-                        }
+                                },
+                            ],
+                        },
                     ],
                     'Valid' => '1'
                 },
@@ -505,6 +505,78 @@ my @Tests = (
             ',
         },
         ExpectedResult => [],
+    },
+    {
+        Description => 'Setting with comments',
+        Config      => {
+            XMLInput => '<?xml version="1.0" encoding="utf-8"?>
+<otrs_config version="2.0" init="Application">
+    <Setting Name="Test" Required="1" Valid="1">
+        <Description Translatable="1">Test 1.</Description>
+        <Navigation>Core::Ticket</Navigation>
+        <Value>
+            <Array>
+                <Item>1</Item>
+#               <Item>2</Item>
+<!--
+                <Item>2</Item>
+                <Item>2</Item>
+-->
+                <Item>2</Item>
+            </Array>
+        </Value>
+    </Setting>
+</otrs_config>
+            ',
+        },
+        ExpectedResult => [
+            {
+                'XMLContentParsed' => {
+                    'Description' => [
+                        {
+                            'Content'      => 'Test 1.',
+                            'Translatable' => '1',
+                        },
+                    ],
+                    'Name'       => 'Test',
+                    'Navigation' => [
+                        {
+                            'Content' => 'Core::Ticket'
+                        },
+                    ],
+                    'Required' => '1',
+                    'Valid'    => '1',
+                    'Value'    => [
+                        {
+                            'Array' => [
+                                {
+                                    'Item' => [
+                                        {
+                                            'Content' => '1',
+                                        },
+                                        {
+                                            'Content' => '2',
+                                        },
+                                    ],
+                                },
+                            ],
+                        },
+                    ],
+                },
+                'XMLContentRaw' => '<Setting Name="Test" Required="1" Valid="1">
+        <Description Translatable="1">Test 1.</Description>
+        <Navigation>Core::Ticket</Navigation>
+        <Value>
+            <Array>
+                <Item>1</Item>
+
+
+                <Item>2</Item>
+            </Array>
+        </Value>
+    </Setting>',
+            },
+        ],
     },
 );